+++ src/pudl/transform/eia.py
@@ -1242,6 +1242,7 @@ def harvested_entity_asset_factory(
         "_core_eia923__generation_fuel",
         "_core_eia923__generation_fuel_nuclear",
         "_core_eia860__generators",
+        "_core_eia860__generators_wind",
         "_core_eia860__ownership",
         "_core_eia860__plants",
         "_core_eia860__utilities",
@@ -1338,5 +1339,6 @@ finished_eia_assets = [
         "core_eia860__assn_yearly_boiler_emissions_control_equipment": "_core_eia860__boiler_emissions_control_equipment_assn",
         "core_eia860__assn_boiler_cooling": "_core_eia860__boiler_cooling",
         "core_eia860__assn_boiler_stack_flue": "_core_eia860__boiler_stack_flue",
+        "core_eia860__yearly_generators_wind": "_core_eia860__generators_wind",
     }.items()
 ]